Refactor accounts reducer into two sub-reducers
